+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/voting/component/VoteAppBar.kt
@@ -1,7 +1,6 @@
 package co.electriccoin.zcash.ui.screen.voting.component
 
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.runtime.Composable